requirements.txt
Django==5.0.3
djangorestframework==3.14.0
psycopg2-binary==2.9.6
django-cors-headers==4.1.0
openpyxl==3.1.2
minio==7.1.15
django-environ
gunicorn==20.1.0
xlsx2html==0.5.0
beautifulsoup4==4.12.3
django-autocomplete-light==3.11.0
pillow==10.3.0
django-rest-passwordreset==1.4.1
django-filter==24.2